Shared test code is two modules, not one include (#68)
test/common.jl was textually included into all four suite modules: four
compiled copies, and three definitions that resolved a name belonging to
whichever module included them (_merge called that module's row,
load_capturing called its check, write_csv took its HEADER as a default).

Split by who needs what. test/fixtures.jl holds the synthetic media and
the ffprobe readers, used by all four suites; test/harness.jl holds the
CSV plumbing, used only by the two gateway suites, which stops the
tracker and end-to-end suites compiling 80 lines they never call.
capturing takes the thunk instead of reaching for check, and each suite
declares DATADIR before including its helpers, so nothing resolves late.

The per-suite half stays per-suite: a factory parameterised by (loader,
header, artifacts) came to more lines than the five each suite spends on
row, write_csv, check, load_capturing and clean, and hid which loader a
check reaches.

Also: probing.jl and fromage.jl each built their own corrupt-video
fixture, one of them from random bytes. Both now call the shared
deterministic generator, which removes the last rand() from the suite.
