+++ test/verifyruns.jl
@@ -2,13 +2,13 @@ module VerifyRunsTests
 
 using Test
 using Fromage: Fromage
+using ..Fixtures, ..Harness
 
-include("common.jl")
-include("VerifyRuns/helpers.jl")
-
-# Shared artifacts and CSV scratch space, built once (ffmpeg is the slow part).
+# Shared artifacts and CSV scratch space, built once for the whole suite. DATADIR comes
+# first: helpers.jl generates into it and closes over it.
 const DATADIR = mktempdir()
-const ART = setup_artifacts(DATADIR)
+
+include("VerifyRuns/helpers.jl")
 
 @testset "VerifyRuns" begin
     include("VerifyRuns/test_input.jl")
